+++ rusk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `Content-Type: application/json` support for `/on/contracts` endpoint
+
 ## [1.2.0] - 2025-03-20
 
 ### Added