+++ rusk/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `Content-Type: application/json` support for `/on/contracts` endpoint
+- Add transaction serialization check
 
 ## [1.2.0] - 2025-03-20