+++ rusk/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add transaction serialization check
 - Add max transaction size check
 
+### Changed
+
+- Upgrade piecrust to `0.28.1`
+
 ## [1.2.0] - 2025-03-20
 
 ### Added