+++ rusk/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2025-04-17
+
 ### Added
 
 - Add `Content-Type: application/json` support for `/on/contracts` endpoint
@@ -406,7 +408,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#290]: https://github.com/dusk-network/rusk/issues/290
 
 
-[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.2.0...HEAD
+[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.3.0...HEAD
+[1.3.0]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.2.0...dusk-rusk-1.3.0
 [1.2.0]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.1.1...dusk-rusk-1.2.0
 [1.1.1]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.1.0...dusk-rusk-1.1.1
 [1.1.0]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.0.2...dusk-rusk-1.1.0